Fix GCC 15 module conflicts and sign-conversion warning in TestRunner
Fixed two GCC 15 issues:
1. Sign-conversion warning in TestRunner.h:104
   - Cast results.size() to long before division

2. Module conflicts when mixing headers with import std
   - Removed "import std;" from test files
   - TestRunner.h provides all needed STL via traditional headers
   - Only import MduX modules (not std module)

GCC 15 has severe bugs with mixing traditional headers and "import std",
causing "conflicting imported declaration" errors. The workaround is to use
traditional headers in files that need both.

All tests build successfully on Windows MSVC and should now work on GCC 15.

+++ tests/TestRunner.h
@@ -101,7 +101,7 @@ public:
 
         std::cout << "\nPerformance:\n";
         std::cout << "  Total time: " << totalTime.count() << " μs\n";
-        std::cout << "  Average:    " << (results.empty() ? 0 : totalTime.count() / results.size()) << " μs/test\n";
+        std::cout << "  Average:    " << (results.empty() ? 0 : totalTime.count() / static_cast<long>(results.size())) << " μs/test\n";
         std::cout << "=============================================================================\n";
     }
